Keep single source of truth for default site path folder
