--- apps/cli/lib/generate-site-name.ts
@@ -1,18 +1,0 @@
-import os from 'os';
-import path from 'path';
-import { generateSiteName as generateSiteNameShared } from '@studio/common/lib/generate-site-name';
-import { sanitizeFolderName } from '@studio/common/lib/sanitize-folder-name';
-import { readAppdata } from 'cli/lib/appdata';
-
-const DEFAULT_SITES_DIR = path.join( os.homedir(), 'Studio' );
-
-export function getDefaultSitePath( siteName: string ): string {
-	const folderName = sanitizeFolderName( siteName );
-	return path.join( DEFAULT_SITES_DIR, folderName );
-}
-
-export async function generateSiteName(): Promise< string > {
-	const appdata = await readAppdata();
-	const usedNames = appdata.sites.map( ( site ) => site.name );
-	return generateSiteNameShared( usedNames, DEFAULT_SITES_DIR );
-}
